add mmap interface for mmap tar file to xrt::bo (#136)
# Summary of Changes

In order to save memory footprint, we need to create an mmap the tar
file.

# Motivation

When memory map a file to a xrt::bo, we don't even need to copy the
constant weights from the tar file to xrt::bo
Not only we save memory usage, but also save the some copying overhead.

# Implementation

to support this feature, we need to add a new interface method
`MetaIOReader::mmap`.

related to JIRA https://jira.xilinx.com/browse/VAI-11102

---------

Co-authored-by: Wang Chunye <chunywan@amd.com>
Co-authored-by: mingyue <mingyue@amd.com>
Co-authored-by: Wang Chunye <chunywan@xilinx.com>