call deinitialize_vaip at dllmain (#172)
There was an intermittent unloading issue which happens at random when
we run the app. Our team had a discussion with Tim if this could be due
to the onnxruntime code added for WinML
In parallel we attempted a fix for this and verified at our end that it
works fine.
Also on the compile failure, we are not able to reproduce the issue, we
suspect this unloading issue could be related
 
unloading_fix_EP2168.zip
 
The fix involves changes in
Onnxruntime. A PR is raised for this
https://github.com/microsoft/win-onnxruntime/pull/105
This PR impacts onnxruntime_providers_vitisai.dll  
Fix in our ep dll onnxruntime_vitisai_ep.dll
 
We built these dlls with the changes and attached both the dlls. You can
use these dlls to check on your end as well. we tested squeeznet and
resnet for 100 iterations in a loop and it's stable.
When you build next time, you can use the PR so that you can get the
changes in onnxruntime_providers_vitisai.dll
Please push this changes into integration


see also db1a5fa

---------

Co-authored-by: Wang Chunye <chunywan@amd.com>