[Proposal] Move remaining model clone to under cache hit check (#259)
<!--
Copyright (C) 2023 - 2025 Advanced Micro Devices, Inc. All rights
reserved.
Licensed under the MIT License.
-->
# Summary of Changes

* The remaining model clone in vaip core is still in the critical run
path. The cloned graph is only used when there is no cache present (like
first time compile) or if the cache needs to be updated during runtime.
Hence if the cache is present, the clone is not needed which would
remove the remaining peak in the memory consumption during inference
time. Proposal is to move that model clone under the case when its
needed.

# Motivation

This is the second model clone that we have discussed recently and
contributes to peak in memory consumption during inference runtime. We
need to get rid of that.

# Implementation

The model clone segment of the code was moved to under the function and
if condition that checks if there was a cache hit.

Co-authored-by: Sachin Kumawat <sachink@amd.com>