[MMAP] Add provider option "enable_mmap" (#266)
# Summary of Changes

There is a feature to create a tar file on a region of memory which is
mapped on the disk file. #136
#252 

This PR adds a provider option "enable_mmap" to control wether to enable
or disable this feature.

# Motivation

"mmap" feature is considered to be risky for some reasons, end-users
would like to disable this feature in case it is not appropriate.

# Implementation

Add a new flag for `TarFile::create_from_path` and pass
`PassContext::get_provider_option("ep_context_enable_mmap")` to it.

---------

Co-authored-by: Wang Chunye <chunywan@amd.com>