enhance environment variable handling for provider options (#368)
# Summary of Changes

import provider options from environment variables.


# Motivation

so we if have a way to override the config and read contents from the
config instead of what is hardcoded in dll, we can debug it quickly

Q: So it is only for debugging purposes?
A: yes, it will be useful for overriding targets, xclbin etc

# Implementation

how about this suggestion? all environment variables which has prefix
with` "VITISAI_EP_PROVIDER_OPTION_<XXX>=YYY"`, makes as if we add a
provider option provider_option["XXX"]="YYY"?

we use `_` as separator because `.` is not a valid separator in bash.

---------

Co-authored-by: Wang Chunye <chunywan@amd.com>