vaip_plugin: check func_set_ is valid before use (#370)
# Summary of Changes

This PR fixes a null pointer dereference whenever a VAIP plugin has
failed to open but the `Plugin` object exists.

# Motivation

There is no way for a `vaip_core::Plugin` to tell whether a `Plugin` is
actually backed by a library, or if that library was not found. The API
lets the user check if symbols exist in the library
(`Plugin::has_method`), that calls `my_plugin_sym` which will then use a
`func_set_`, known to be `nullptr` if the library was not loaded. In
that case, we have `my_plugin_sym` return `nullptr` directly, signaling
to `has_method` that the requested symbol is not available.

# Implementation

There is nothing noteworthy implementation-wise.