Allow to define different kernels for a Custom Op (#390)
<!--
Copyright (C) 2023 - 2025 Advanced Micro Devices, Inc. All rights
reserved.
Licensed under the MIT License.
-->
# Summary of Changes

Adding functionality to inject a different kernel to `XilinxCustomOp`.
This makes the class more flexible and allows us to error out if the
`Compute` function of the kernel is called. The default implementation
is to print a message to the log.

# Motivation

This solves a problem where the custom op would not get compiled to the
AIE and instead try to execute in the CPU. While doing so, it would
print a message to the log and the result of `onnx_session.run` would be
`[None]`. We're using the functionality of this PR to hard fail if the
`Compute` function is ever invoked which, for us, only happens if the
operator is not mapped to the AIE.

# Implementation

Using templates to achieve some type of static polymorphism